+++ espflash/src/cli/mod.rs
@@ -373,6 +373,13 @@ pub fn board_info(args: &ConnectArgs, config: &Config) -> Result<()> {
     let mut flasher = connect(args, config, true, true)?;
     print_board_info(&mut flasher)?;
 
+    if flasher.chip() != Chip::Esp32 {
+        let security_info = flasher.get_security_info()?;
+        println!("{security_info}");
+    } else {
+        println!("Security features: None");
+    }
+
     Ok(())
 }